Refactor and clean up various SearchForm things
