+++ de-lib/src/main/java/org/iplantc/de/client/models/search/Fields.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.client.models.querydsl;
+package org.iplantc.de.client.models.search;
 
 import org.iplantc.de.client.models.sharing.PermissionValue;
