+++ de-lib/src/main/java/org/iplantc/de/client/models/search/Metadata.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.client.models.querydsl;
+package org.iplantc.de.client.models.search;
 
 /**
  * The autobean representation of the metadata returned in an Elasticsearch query response
