+++ de-lib/src/main/java/org/iplantc/de/client/models/search/MetadataList.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.client.models.querydsl;
+package org.iplantc.de.client.models.search;
 
 import java.util.List;
