+++ de-lib/src/main/java/org/iplantc/de/client/models/search/SearchAutoBeanFactory.java
@@ -1,5 +1,8 @@
 package org.iplantc.de.client.models.search;
 
+import org.iplantc.de.client.models.diskResources.DiskResource;
+import org.iplantc.de.client.models.diskResources.File;
+import org.iplantc.de.client.models.diskResources.Folder;
 import org.iplantc.de.client.models.search.FileSizeRange.FileSizeUnit;
 
 import com.google.gwt.core.client.GWT;
@@ -19,5 +22,27 @@ public interface SearchAutoBeanFactory extends AutoBeanFactory {
     AutoBean<FileSizeUnit> fileSizeUnit();
 
     AutoBean<DiskResourceQueryTemplateList> drQtList();
+
+    AutoBean<SearchResponse> getSearchResponse();
+
+    AutoBean<DiskResource> getDiskResource();
+
+    AutoBean<File> getFile();
+
+    AutoBean<Folder> getFolder();
+
+    AutoBean<DocumentList> getDocumentList();
+
+    AutoBean<Document> getDocument();
+
+    AutoBean<Metadata> getMetadata();
+
+    AutoBean<MetadataList> getMetadataList();
+
+    AutoBean<Source> getSource();
+
+    AutoBean<DocumentUserPermissionList> getDocumentUserPermissionList();
+
+    AutoBean<DiskResourceQueryTemplate> getQueryTemplate();
 
 }
