+++ de-lib/src/main/java/org/iplantc/de/client/models/search/SearchResponse.java
@@ -1,4 +1,6 @@
-package org.iplantc.de.client.models.querydsl;
+package org.iplantc.de.client.models.search;
+
+import org.iplantc.de.client.models.search.Document;
 
 import com.google.web.bindery.autobean.shared.AutoBean.PropertyName;
