+++ de-lib/src/main/java/org/iplantc/de/client/models/search/Source.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.client.models.querydsl;
+package org.iplantc.de.client.models.search;
 
 import org.iplantc.de.client.models.HasId;
 import org.iplantc.de.client.models.sharing.OldUserPermission;
