+++ de-lib/src/main/java/org/iplantc/de/client/util/SearchModelUtils.java
@@ -1,6 +1,5 @@
 package org.iplantc.de.client.util;
 
-import org.iplantc.de.client.models.querydsl.QueryAutoBeanFactory;
 import org.iplantc.de.client.models.search.DiskResourceQueryTemplate;
 import org.iplantc.de.client.models.search.FileSizeRange.FileSizeUnit;
 import org.iplantc.de.client.models.search.SearchAutoBeanFactory;
@@ -26,7 +25,6 @@ public class SearchModelUtils {
 
     private final List<String> fileSizeUnits = Lists.newArrayList("KB", "MB", "GB", "TB");
     private final SearchAutoBeanFactory factory = GWT.create(SearchAutoBeanFactory.class);
-    private final QueryAutoBeanFactory queryFactory = GWT.create(QueryAutoBeanFactory.class);
     private static SearchModelUtils INSTANCE;
     Logger LOG = Logger.getLogger(SearchModelUtils.class.getName());
