+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/presenters/navigation/proxy/FolderRpcProxyImpl.java
@@ -8,6 +8,7 @@ import org.iplantc.de.client.models.diskResources.RootFolders;
 import org.iplantc.de.client.models.search.DiskResourceQueryTemplate;
 import org.iplantc.de.client.services.DiskResourceServiceFacade;
 import org.iplantc.de.client.services.SearchServiceFacade;
+import org.iplantc.de.client.util.SearchModelUtils;
 import org.iplantc.de.commons.client.ErrorHandler;
 import org.iplantc.de.commons.client.info.ErrorAnnouncementConfig;
 import org.iplantc.de.commons.client.info.IplantAnnouncer;
@@ -27,13 +28,9 @@ import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.safehtml.shared.SafeHtmlUtils;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.inject.Inject;
-import com.google.web.bindery.autobean.shared.AutoBeanCodex;
-import com.google.web.bindery.autobean.shared.AutoBeanUtils;
 import com.google.web.bindery.autobean.shared.Splittable;
 
 import com.sencha.gxt.data.client.loader.RpcProxy;
-
-import org.springframework.beans.annotation.AnnotationBeanUtils;
 
 import java.util.Collections;
 import java.util.List;
@@ -155,6 +152,7 @@ public class FolderRpcProxyImpl extends RpcProxy<Folder, List<Folder>> implement
     private final SearchServiceFacade searchService;
     private IsMaskable isMaskable;
     @Inject UserInfo userInfo;
+    @Inject SearchModelUtils searchModelUtils;
 
     @Inject
     FolderRpcProxyImpl(final DiskResourceServiceFacade drService,
@@ -202,7 +200,7 @@ public class FolderRpcProxyImpl extends RpcProxy<Folder, List<Folder>> implement
             }
         } else if (parentFolder instanceof DiskResourceQueryTemplate) {
             DiskResourceQueryTemplate template = (DiskResourceQueryTemplate)parentFolder;
-            Splittable encode = AutoBeanCodex.encode(AutoBeanUtils.getAutoBean(template));
+            Splittable encode = searchModelUtils.convertTemplateToSplittable(template);
             fireEvent(new SubmitDiskResourceQueryEvent(encode));
         } else {
             drService.getSubFolders(parentFolder, new SubFoldersCallback(callback,
