--- de-lib/src/main/java/org/iplantc/de/diskResource/client/views/search/SearchViewImpl.ui.xml
@@ -1,153 +1,0 @@
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder'
-             xmlns:container="urn:import:com.sencha.gxt.widget.core.client.container"
-             xmlns:form="urn:import:com.sencha.gxt.widget.core.client.form"
-             xmlns:grid="urn:import:com.sencha.gxt.widget.core.client.grid"
-             xmlns:button="urn:import:com.sencha.gxt.widget.core.client.button"
-             xmlns:userSearch="urn:import:org.iplantc.de.collaborators.client.util"
->
-
-    <ui:with field="appearance"
-             type="org.iplantc.de.diskResource.client.SearchView.SearchViewAppearance">
-    </ui:with>
-
-    <ui:with type="com.sencha.gxt.widget.core.client.container.CssFloatLayoutContainer.CssFloatData"
-             field="leftColumn">
-        <ui:attributes size="0.5" margins="{leftMargins}"/>
-    </ui:with>
-
-    <ui:with type="com.sencha.gxt.widget.core.client.container.CssFloatLayoutContainer.CssFloatData"
-             field="rightColumn">
-        <ui:attributes size="0.5" margins="{rightMargins}"/>
-    </ui:with>
-
-    <ui:with field="leftMargins"
-             type="com.sencha.gxt.core.client.util.Margins">
-        <ui:attributes top="5" right="8" bottom="0" left="5"/>
-    </ui:with>
-
-    <ui:with field="rightMargins"
-             type="com.sencha.gxt.core.client.util.Margins">
-        <ui:attributes top="5" right="5" bottom="0" left="8"/>
-    </ui:with>
-
-    <ui:with field="verticalLayout"
-             type="com.sencha.gxt.widget.core.client.container.VerticalLayoutContainer.VerticalLayoutData">
-        <ui:attributes width="-1" height="-1"/>
-    </ui:with>
-
-    <ui:with field="verticalMargins"
-             type="com.sencha.gxt.core.client.util.Margins">
-        <ui:attributes top="15" right="15" bottom="0" left="15"/>
-    </ui:with>
-
-    <ui:with field="permissionUsers" type="com.sencha.gxt.data.shared.ListStore" />
-    <ui:with field="cm"
-             type="com.sencha.gxt.widget.core.client.grid.ColumnModel" />
-    <ui:with field="gridView"
-             type="com.sencha.gxt.widget.core.client.grid.GridView">
-        <ui:attributes forceFit="true" autoFill="true" />
-    </ui:with>
-
-    <container:VerticalLayoutContainer layoutData="{verticalLayout}"
-                                       ui:field="con"
-                                       width="600px"
-                                       borders="true">
-        <container:child>
-            <container:CssFloatLayoutContainer>
-                <container:child layoutData="{leftColumn}">
-                    <form:FieldLabel text="{appearance.nameHas}"
-                                     labelAlign="TOP">
-                        <form:widget>
-                            <form:TextField ui:field="nameHas"
-                                            emptyText="{appearance.emptyText}"/>
-                        </form:widget>
-                    </form:FieldLabel>
-                </container:child>
-                <container:child layoutData="{rightColumn}">
-                    <form:CheckBox ui:field="exactNameMatch"
-                                   boxLabel="{appearance.exactNameMatch}"/>
-                </container:child>
-            </container:CssFloatLayoutContainer>
-        </container:child>
-
-        <container:child>
-            <container:CssFloatLayoutContainer>
-                <container:child layoutData="{leftColumn}">
-                    <form:FieldLabel text="{appearance.pathPrefix}"
-                                     labelAlign="TOP">
-                        <form:widget>
-                            <form:TextField ui:field="pathPrefix"
-                                            emptyText="{appearance.emptyText}"/>
-                        </form:widget>
-                    </form:FieldLabel>
-                </container:child>
-                <container:child layoutData="{rightColumn}">
-                    <form:FieldLabel text="{appearance.owner}"
-                                     labelAlign="TOP">
-                        <form:widget>
-                            <form:TextField ui:field="owner"
-                                            emptyText="{appearance.emptyText}"/>
-                        </form:widget>
-                    </form:FieldLabel>
-                </container:child>
-            </container:CssFloatLayoutContainer>
-        </container:child>
-
-        <container:child>
-            <container:CssFloatLayoutContainer>
-                <container:child layoutData="{leftColumn}">
-                    <container:VerticalLayoutContainer>
-                        <container:child>
-                            <form:FieldLabel text="{appearance.sharedWith}"
-                                             labelAlign="TOP">
-                            </form:FieldLabel>
-                        </container:child>
-                        <container:child>
-                            <userSearch:UserSearchField ui:field="userSearchField"/>
-                        </container:child>
-                        <container:child>
-                            <grid:Grid ui:field="usersGrid"
-                                       cm="{cm}"
-                                       store="{permissionUsers}"
-                                       view="{gridView}"
-                                       loadMask="true"
-                                       columnReordering="true"
-                                       borders="true"
-                                       height="100px">
-                            </grid:Grid>
-                        </container:child>
-                    </container:VerticalLayoutContainer>
-                </container:child>
-
-                <container:child layoutData="{rightColumn}">
-                    <form:FieldLabel text="{appearance.permissionValueLabel}"
-                                     labelAlign="TOP">
-                        <form:widget>
-                            <form:StringComboBox ui:field="permission"
-                                                 triggerAction="ALL"
-                                                 editable="false"/>
-                        </form:widget>
-                    </form:FieldLabel>
-                </container:child>
-                <container:child layoutData="{rightColumn}">
-                    <form:CheckBox ui:field="permissionRecurse"
-                                   boxLabel="{appearance.permissionRecurse}"/>
-                </container:child>
-            </container:CssFloatLayoutContainer>
-        </container:child>
-
-        <container:child>
-            <container:CssFloatLayoutContainer>
-                <container:child layoutData="{leftColumn}">
-                    <button:TextButton height="0px"/>
-                </container:child>
-                <container:child layoutData="{rightColumn}">
-                    <button:TextButton ui:field="searchBtn"
-                                       text="{appearance.searchBtn}"/>
-                </container:child>
-            </container:CssFloatLayoutContainer>
-        </container:child>
-
-    </container:VerticalLayoutContainer>
-</ui:UiBinder>
