--- de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/cells/DeleteCell.gss
@@ -1,3 +1,0 @@
-.delete {
-	cursor: pointer;
-}
