+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/diskResource/search/SearchMessages.java
@@ -12,86 +12,4 @@ public interface SearchMessages extends Messages {
     String deleteSearchSuccess(String searchName);
 
     String saveQueryTemplateFail();
-
-    String nameHas();
-
-    String emptyText();
-
-    String nameHasNot();
-
-    String createdWithin();
-
-    String modifiedWithin();
-
-    String metadataAttributeHas();
-
-    String metadataValueHas();
-
-    String ownedBy();
-
-    String emptyNameText();
-
-    String sharedWith();
-
-    String fileSizeGreaterThan();
-
-    String fileSizeLessThan();
-
-    String includeTrash();
-
-    String searchBtnText();
-
-    String emptyTimeText();
-
-    String fileNameClass();
-
-    String createdWithinClass();
-
-    String fileNameNegateClass();
-
-    String modifiedWithinClass();
-
-    String metadataAttributeClass();
-
-    String ownerClass();
-
-    String metadataValueClass();
-
-    String sharedClass();
-
-    String searchClass();
-
-    String fileSizeClass();
-
-    String tagsClass();
-
-    String trashAndFilterClass();
-
-    String pathPrefix();
-
-    String exactNameMatch();
-
-    String owner();
-
-    String exactUserNameMatch();
-
-    String permissionValue();
-
-    String permissionRecurse();
-
-    String permissionUsers();
-
-    String emptyDropDownText();
-
-    String taggedWith();
-
-    String fileSizes();
-
-    String saveSearchBtnText();
-
-    String filterName();
-
-    String requiredField();
-
-    String saveSearchTitle();
 }
