+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/diskResource/search/SearchMessages.properties
@@ -1,44 +1,3 @@
 advancedSearchToolTip = Click here for advanced search options
-createdWithin=Created within
-createdWithinClass=createdwithin
 deleteSearchSuccess=Successfully deleted saved search: {0}
-emptyDropDownText=---
-emptyNameText=Enter CyVerse user name
-emptyText=Enter values...
-emptyTimeText=---
-exactNameMatch=Exact Name Match
-exactUserNameMatch=Exact Username Match
-fileNameClass=filename
-fileNameNegateClass=negatefilename
-fileSizeClass=filesize
-fileSizeGreaterThan=File size is bigger than or equal to
-fileSizeLessThan=File size is smaller than or equal to
-fileSizes=KB, MB, GB, TB
-filterName=Filter Name
-includeTrash=Include items in Trash
-metadataAttributeClass=metadataattrib
-metadataAttributeHas=Metadata attribute has the words
-metadataValueClass=metadataval
-metadataValueHas=Metadata value has the words
-modifiedWithin=Modified within
-modifiedWithinClass=modifiedwithin
-nameHas=File/Folder name has the words
-nameHasNot=File/Folder name doesn''t have
-ownedBy=Owned by
-owner=Owner
-ownerClass=owner
-pathPrefix=Path Prefix
-permissionRecurse=Permission Recurse
-permissionUsers=Permission Users
-permissionValue=Permission Value
-requiredField=This field is required
 saveQueryTemplateFail=Unable to save filter.
-saveSearchBtnText=Save Search
-saveSearchTitle=Save Filter
-searchBtnText=Search
-searchClass=search
-sharedClass=shared
-sharedWith=Shared with
-taggedWith=Tagged with
-tagsClass=tags
-trashAndFilterClass=trashandfilter
